rename customer account and shipping info
