+++ src/components/TheHeader/template.html
@@ -30,7 +30,7 @@
           <router-link
             v-if="customer.id"
             class="nav-link"
-            :to="{ name: 'customer_info' }">{{ username }}</router-link>
+            :to="{ name: 'customer-account' }">{{ username }}</router-link>
 
         </li>
         <li class="nav-item">
