+++ src/router_views/Customer/template.html
@@ -4,7 +4,7 @@
 
       <ul class="nav nav-pills flex-column">
         <li class="nav-item">
-          <router-link class="nav-link" active-class="active" :to="{ name: 'customer_info' }">
+          <router-link class="nav-link" active-class="active" :to="{ name: 'customer-account' }">
             {{ $t('customer.account') }}
           </router-link>
         </li>
@@ -15,7 +15,7 @@
           <a class="nav-link" href="#">{{ $t('customer.payment_info') }}</a>
         </li>
         <li class="nav-item">
-          <router-link class="nav-link" active-class="active" :to="{ name: 'customer_addresses' }">
+          <router-link class="nav-link" active-class="active" :to="{ name: 'customer-shipping-info' }">
             {{ $t('customer.shipping_info') }}
           </router-link>
         </li>
